Implement review suggestions
