+++ alacritty/src/input.rs
@@ -31,8 +31,8 @@ use alacritty_terminal::vi_mode::ViMotion;
 use crate::clipboard::Clipboard;
 use crate::config::{Action, BindingMode, Config, Key, SearchAction, ViAction};
 use crate::daemon::start_daemon;
+use crate::display::hint::HintState;
 use crate::display::window::Window;
-use crate::display::HintState;
 use crate::event::{ClickState, Event, Mouse, TYPING_SEARCH_DELAY};
 use crate::message_bar::{self, Message};
 use crate::scheduler::{Scheduler, TimerId};
@@ -115,226 +115,227 @@ pub trait ActionContext<T: EventListener> {
     fn hint_input(&mut self, _character: char) {}
 }
 
-macro_rules! execute_action {
-    ($ctx:expr, $action:expr) => {
-        #[allow(clippy::unnecessary_mut_passed)]
-        match $action {
+trait Execute<T: EventListener> {
+    fn execute<A: ActionContext<T>>(&self, ctx: &mut A);
+}
+
+impl<T: EventListener> Execute<T> for Action {
+    #[inline]
+    fn execute<A: ActionContext<T>>(&self, ctx: &mut A) {
+        match self {
             Action::Esc(s) => {
                 let bytes = s.clone().into_bytes();
 
-                $ctx.on_typing_start();
+                ctx.on_typing_start();
 
-                $ctx.clear_selection();
-                $ctx.scroll(Scroll::Bottom);
-                $ctx.write_to_pty(bytes)
+                ctx.clear_selection();
+                ctx.scroll(Scroll::Bottom);
+                ctx.write_to_pty(bytes)
             },
             Action::Command(program) => start_daemon(program.program(), program.args()),
             Action::Hint(hint) => {
-                let hint = hint.clone();
-                $ctx.hint_state().start(hint);
-                $ctx.mark_dirty();
+                ctx.hint_state().start(hint.clone());
+                ctx.mark_dirty();
             },
-            Action::ToggleViMode => $ctx.toggle_vi_mode(),
+            Action::ToggleViMode => ctx.toggle_vi_mode(),
             Action::ViMotion(motion) => {
-                let motion = *motion;
-                $ctx.on_typing_start();
-                $ctx.terminal_mut().vi_motion(motion);
-                $ctx.mark_dirty();
+                ctx.on_typing_start();
+                ctx.terminal_mut().vi_motion(*motion);
+                ctx.mark_dirty();
             },
             Action::ViAction(ViAction::ToggleNormalSelection) => {
-                $ctx.toggle_vi_selection(SelectionType::Simple);
+                ctx.toggle_vi_selection(SelectionType::Simple);
             },
             Action::ViAction(ViAction::ToggleLineSelection) => {
-                $ctx.toggle_vi_selection(SelectionType::Lines);
+                ctx.toggle_vi_selection(SelectionType::Lines);
             },
             Action::ViAction(ViAction::ToggleBlockSelection) => {
-                $ctx.toggle_vi_selection(SelectionType::Block);
+                ctx.toggle_vi_selection(SelectionType::Block);
             },
             Action::ViAction(ViAction::ToggleSemanticSelection) => {
-                $ctx.toggle_vi_selection(SelectionType::Semantic);
+                ctx.toggle_vi_selection(SelectionType::Semantic);
             },
             Action::ViAction(ViAction::Open) => {
-                $ctx.mouse_mut().block_url_launcher = false;
-                let point = $ctx.terminal().vi_mode_cursor.point;
-                if let Some(url) = $ctx.urls().find_at(point) {
-                    $ctx.launch_url(url);
+                ctx.mouse_mut().block_url_launcher = false;
+                let point = ctx.terminal().vi_mode_cursor.point;
+                if let Some(url) = ctx.urls().find_at(point) {
+                    ctx.launch_url(url);
                 }
             },
             Action::ViAction(ViAction::SearchNext) => {
-                let direction = $ctx.search_direction();
-                let terminal = $ctx.terminal();
+                let direction = ctx.search_direction();
+                let terminal = ctx.terminal();
                 let vi_point = terminal.visible_to_buffer(terminal.vi_mode_cursor.point);
                 let origin = match direction {
                     Direction::Right => vi_point.add_absolute(terminal, Boundary::Wrap, 1),
                     Direction::Left => vi_point.sub_absolute(terminal, Boundary::Wrap, 1),
                 };
 
-                if let Some(regex_match) = $ctx.search_next(origin, direction, Side::Left) {
-                    $ctx.terminal_mut().vi_goto_point(*regex_match.start());
-                    $ctx.mark_dirty();
+                if let Some(regex_match) = ctx.search_next(origin, direction, Side::Left) {
+                    ctx.terminal_mut().vi_goto_point(*regex_match.start());
+                    ctx.mark_dirty();
                 }
             },
             Action::ViAction(ViAction::SearchPrevious) => {
-                let direction = $ctx.search_direction().opposite();
-                let terminal = $ctx.terminal();
+                let direction = ctx.search_direction().opposite();
+                let terminal = ctx.terminal();
                 let vi_point = terminal.visible_to_buffer(terminal.vi_mode_cursor.point);
                 let origin = match direction {
                     Direction::Right => vi_point.add_absolute(terminal, Boundary::Wrap, 1),
                     Direction::Left => vi_point.sub_absolute(terminal, Boundary::Wrap, 1),
                 };
 
-                if let Some(regex_match) = $ctx.search_next(origin, direction, Side::Left) {
-                    $ctx.terminal_mut().vi_goto_point(*regex_match.start());
-                    $ctx.mark_dirty();
+                if let Some(regex_match) = ctx.search_next(origin, direction, Side::Left) {
+                    ctx.terminal_mut().vi_goto_point(*regex_match.start());
+                    ctx.mark_dirty();
                 }
             },
             Action::ViAction(ViAction::SearchStart) => {
-                let terminal = $ctx.terminal();
+                let terminal = ctx.terminal();
                 let origin = terminal
                     .visible_to_buffer(terminal.vi_mode_cursor.point)
                     .sub_absolute(terminal, Boundary::Wrap, 1);
 
-                if let Some(regex_match) = $ctx.search_next(origin, Direction::Left, Side::Left) {
-                    $ctx.terminal_mut().vi_goto_point(*regex_match.start());
-                    $ctx.mark_dirty();
+                if let Some(regex_match) = ctx.search_next(origin, Direction::Left, Side::Left) {
+                    ctx.terminal_mut().vi_goto_point(*regex_match.start());
+                    ctx.mark_dirty();
                 }
             },
             Action::ViAction(ViAction::SearchEnd) => {
-                let terminal = $ctx.terminal();
+                let terminal = ctx.terminal();
                 let origin = terminal
                     .visible_to_buffer(terminal.vi_mode_cursor.point)
                     .add_absolute(terminal, Boundary::Wrap, 1);
 
-                if let Some(regex_match) = $ctx.search_next(origin, Direction::Right, Side::Right) {
-                    $ctx.terminal_mut().vi_goto_point(*regex_match.end());
-                    $ctx.mark_dirty();
+                if let Some(regex_match) = ctx.search_next(origin, Direction::Right, Side::Right) {
+                    ctx.terminal_mut().vi_goto_point(*regex_match.end());
+                    ctx.mark_dirty();
                 }
             },
             Action::SearchAction(SearchAction::SearchFocusNext) => {
-                let direction = $ctx.search_direction();
-                $ctx.advance_search_origin(direction);
+                ctx.advance_search_origin(ctx.search_direction());
             },
             Action::SearchAction(SearchAction::SearchFocusPrevious) => {
-                let direction = $ctx.search_direction().opposite();
-                $ctx.advance_search_origin(direction);
+                let direction = ctx.search_direction().opposite();
+                ctx.advance_search_origin(direction);
             },
-            Action::SearchAction(SearchAction::SearchConfirm) => $ctx.confirm_search(),
-            Action::SearchAction(SearchAction::SearchCancel) => $ctx.cancel_search(),
+            Action::SearchAction(SearchAction::SearchConfirm) => ctx.confirm_search(),
+            Action::SearchAction(SearchAction::SearchCancel) => ctx.cancel_search(),
             Action::SearchAction(SearchAction::SearchClear) => {
-                let direction = $ctx.search_direction();
-                $ctx.cancel_search();
-                $ctx.start_search(direction);
+                let direction = ctx.search_direction();
+                ctx.cancel_search();
+                ctx.start_search(direction);
             },
-            Action::SearchAction(SearchAction::SearchDeleteWord) => $ctx.search_pop_word(),
+            Action::SearchAction(SearchAction::SearchDeleteWord) => ctx.search_pop_word(),
             Action::SearchAction(SearchAction::SearchHistoryPrevious) => {
-                $ctx.search_history_previous()
+                ctx.search_history_previous()
             },
-            Action::SearchAction(SearchAction::SearchHistoryNext) => $ctx.search_history_next(),
-            Action::SearchForward => $ctx.start_search(Direction::Right),
-            Action::SearchBackward => $ctx.start_search(Direction::Left),
-            Action::Copy => $ctx.copy_selection(ClipboardType::Clipboard),
+            Action::SearchAction(SearchAction::SearchHistoryNext) => ctx.search_history_next(),
+            Action::SearchForward => ctx.start_search(Direction::Right),
+            Action::SearchBackward => ctx.start_search(Direction::Left),
+            Action::Copy => ctx.copy_selection(ClipboardType::Clipboard),
             #[cfg(not(any(target_os = "macos", windows)))]
-            Action::CopySelection => $ctx.copy_selection(ClipboardType::Selection),
-            Action::ClearSelection => $ctx.clear_selection(),
+            Action::CopySelection => ctx.copy_selection(ClipboardType::Selection),
+            Action::ClearSelection => ctx.clear_selection(),
             Action::Paste => {
-                let text = $ctx.clipboard_mut().load(ClipboardType::Clipboard);
-                paste($ctx, &text);
+                let text = ctx.clipboard_mut().load(ClipboardType::Clipboard);
+                paste(ctx, &text);
             },
             Action::PasteSelection => {
-                let text = $ctx.clipboard_mut().load(ClipboardType::Selection);
-                paste($ctx, &text);
+                let text = ctx.clipboard_mut().load(ClipboardType::Selection);
+                paste(ctx, &text);
             },
-            Action::ToggleFullscreen => $ctx.window_mut().toggle_fullscreen(),
+            Action::ToggleFullscreen => ctx.window_mut().toggle_fullscreen(),
             #[cfg(target_os = "macos")]
-            Action::ToggleSimpleFullscreen => $ctx.window_mut().toggle_simple_fullscreen(),
+            Action::ToggleSimpleFullscreen => ctx.window_mut().toggle_simple_fullscreen(),
             #[cfg(target_os = "macos")]
-            Action::Hide => $ctx.event_loop().hide_application(),
+            Action::Hide => ctx.event_loop().hide_application(),
             #[cfg(not(target_os = "macos"))]
-            Action::Hide => $ctx.window().set_visible(false),
-            Action::Minimize => $ctx.window().set_minimized(true),
-            Action::Quit => $ctx.terminal_mut().exit(),
-            Action::IncreaseFontSize => $ctx.change_font_size(FONT_SIZE_STEP),
-            Action::DecreaseFontSize => $ctx.change_font_size(FONT_SIZE_STEP * -1.),
-            Action::ResetFontSize => $ctx.reset_font_size(),
+            Action::Hide => ctx.window().set_visible(false),
+            Action::Minimize => ctx.window().set_minimized(true),
+            Action::Quit => ctx.terminal_mut().exit(),
+            Action::IncreaseFontSize => ctx.change_font_size(FONT_SIZE_STEP),
+            Action::DecreaseFontSize => ctx.change_font_size(FONT_SIZE_STEP * -1.),
+            Action::ResetFontSize => ctx.reset_font_size(),
             Action::ScrollPageUp => {
                 // Move vi mode cursor.
-                let term = $ctx.terminal_mut();
+                let term = ctx.terminal_mut();
                 let scroll_lines = term.screen_lines().0 as isize;
                 term.vi_mode_cursor = term.vi_mode_cursor.scroll(term, scroll_lines);
 
-                $ctx.scroll(Scroll::PageUp);
+                ctx.scroll(Scroll::PageUp);
             },
             Action::ScrollPageDown => {
                 // Move vi mode cursor.
-                let term = $ctx.terminal_mut();
+                let term = ctx.terminal_mut();
                 let scroll_lines = -(term.screen_lines().0 as isize);
                 term.vi_mode_cursor = term.vi_mode_cursor.scroll(term, scroll_lines);
 
-                $ctx.scroll(Scroll::PageDown);
+                ctx.scroll(Scroll::PageDown);
             },
             Action::ScrollHalfPageUp => {
                 // Move vi mode cursor.
-                let term = $ctx.terminal_mut();
+                let term = ctx.terminal_mut();
                 let scroll_lines = term.screen_lines().0 as isize / 2;
                 term.vi_mode_cursor = term.vi_mode_cursor.scroll(term, scroll_lines);
 
-                $ctx.scroll(Scroll::Delta(scroll_lines));
+                ctx.scroll(Scroll::Delta(scroll_lines));
             },
             Action::ScrollHalfPageDown => {
                 // Move vi mode cursor.
-                let term = $ctx.terminal_mut();
+                let term = ctx.terminal_mut();
                 let scroll_lines = -(term.screen_lines().0 as isize / 2);
                 term.vi_mode_cursor = term.vi_mode_cursor.scroll(term, scroll_lines);
 
-                $ctx.scroll(Scroll::Delta(scroll_lines));
+                ctx.scroll(Scroll::Delta(scroll_lines));
             },
             Action::ScrollLineUp => {
                 // Move vi mode cursor.
-                let term = $ctx.terminal();
+                let term = ctx.terminal();
                 if term.grid().display_offset() != term.history_size()
                     && term.vi_mode_cursor.point.line + 1 != term.screen_lines()
                 {
-                    $ctx.terminal_mut().vi_mode_cursor.point.line += 1;
+                    ctx.terminal_mut().vi_mode_cursor.point.line += 1;
                 }
 
-                $ctx.scroll(Scroll::Delta(1));
+                ctx.scroll(Scroll::Delta(1));
             },
             Action::ScrollLineDown => {
                 // Move vi mode cursor.
-                if $ctx.terminal().grid().display_offset() != 0
-                    && $ctx.terminal().vi_mode_cursor.point.line.0 != 0
+                if ctx.terminal().grid().display_offset() != 0
+                    && ctx.terminal().vi_mode_cursor.point.line.0 != 0
                 {
-                    $ctx.terminal_mut().vi_mode_cursor.point.line -= 1;
+                    ctx.terminal_mut().vi_mode_cursor.point.line -= 1;
                 }
 
-                $ctx.scroll(Scroll::Delta(-1));
+                ctx.scroll(Scroll::Delta(-1));
             },
             Action::ScrollToTop => {
-                $ctx.scroll(Scroll::Top);
+                ctx.scroll(Scroll::Top);
 
                 // Move vi mode cursor.
-                $ctx.terminal_mut().vi_mode_cursor.point.line = Line(0);
-                $ctx.terminal_mut().vi_motion(ViMotion::FirstOccupied);
-                $ctx.mark_dirty();
+                ctx.terminal_mut().vi_mode_cursor.point.line = Line(0);
+                ctx.terminal_mut().vi_motion(ViMotion::FirstOccupied);
+                ctx.mark_dirty();
             },
             Action::ScrollToBottom => {
-                $ctx.scroll(Scroll::Bottom);
+                ctx.scroll(Scroll::Bottom);
 
                 // Move vi mode cursor.
-                let term = $ctx.terminal_mut();
+                let term = ctx.terminal_mut();
                 term.vi_mode_cursor.point.line = term.screen_lines() - 1;
 
                 // Move to beginning twice, to always jump across linewraps.
                 term.vi_motion(ViMotion::FirstOccupied);
                 term.vi_motion(ViMotion::FirstOccupied);
-                $ctx.mark_dirty();
+                ctx.mark_dirty();
             },
-            Action::ClearHistory => $ctx.terminal_mut().clear_screen(ClearMode::Saved),
-            Action::ClearLogNotice => $ctx.pop_message(),
-            Action::SpawnNewInstance => $ctx.spawn_new_instance(),
+            Action::ClearHistory => ctx.terminal_mut().clear_screen(ClearMode::Saved),
+            Action::ClearLogNotice => ctx.pop_message(),
+            Action::SpawnNewInstance => ctx.spawn_new_instance(),
             Action::ReceiveChar | Action::None => (),
         }
-    };
+    }
 }
 
 fn paste<T: EventListener, A: ActionContext<T>>(ctx: &mut A, contents: &str) {
@@ -924,7 +925,7 @@ impl<T: EventListener, A: ActionContext<T>> Processor<T, A> {
                 *suppress_chars.get_or_insert(true) &= binding.action != Action::ReceiveChar;
 
                 // Binding was triggered; run the action.
-                execute_action!(&mut self.ctx, &binding.action);
+                binding.action.clone().execute(&mut self.ctx);
             }
         }
 
@@ -950,7 +951,7 @@ impl<T: EventListener, A: ActionContext<T>> Processor<T, A> {
             }
 
             if binding.is_triggered_by(mode, mods, &button) {
-                execute_action!(&mut self.ctx, &binding.action);
+                binding.action.execute(&mut self.ctx);
             }
         }
     }
